Add disabling editing invoice fields when signed FR
