+++ register/api/registration/cr.py
@@ -1,13 +1,11 @@
 from rest_framework.parsers import JSONParser
 
 from django.http import JsonResponse
-from django.views.decorators.http import require_POST, require_GET
+from django.views.decorators.http import require_POST
 
-from register.api import status
-from register.api.registration.types import (
-    Registration,
-    get_contract_signature,
-)
+from register.api import get_user, status
+from register.api.registration.types.registration import Registration
+from register.api.registration.util import get_contract_signature
 
 from register.models import SignupLog
 from accounting.models import Order
@@ -16,11 +14,15 @@ from accounting.models import Order
 def put_cr_registration(request, registration, purchasing_company):
     data = JSONParser().parse(request)
     serializer = registration.get_serializer(
-        data, context={"purchasing_company": purchasing_company}
+        context={"user": get_user(request), "purchasing_company": purchasing_company},
+        data=data,
     )
 
     if serializer.is_valid():
-        serializer.update(registration, serializer.validated_data)
+        serializer.update(
+            registration,
+            serializer.validated_data
+        )
         return JsonResponse(serializer.data)
 
     return status.serializer_error(serializer.errors)
@@ -49,7 +51,7 @@ def get_registration(company, fair, contact, exhibitor):
 
 def handle_cr(request, company, fair, contact, exhibitor):
     registration = get_registration(company, fair, contact, exhibitor)
-    serializer = registration.get_serializer()
+    serializer = registration.get_serializer(context={"user": get_user(request)})
 
     if request.method == "GET":
         return JsonResponse(serializer.data, safe=False)
@@ -82,9 +84,6 @@ def submit_cr(request, company, fair, contact, exhibitor):
 
     registration = get_registration(company, fair, contact, exhibitor)
 
-    try:
-        serializer = registration.get_serializer()
-    except:
-        return
+    serializer = registration.get_serializer(context={"user": get_user(request)})
 
     return JsonResponse(serializer.data, safe=False)
